+++ src/pkgs/yq-linux.ps1
@@ -1,0 +1,3 @@
+<# Toolchains | SPDX-License-Identifier: MPL-2.0 #>
+Initialize-TlcGitHubCliPackage -Name 'yq-linux' -CanonicalName 'yq' -Owner 'mikefarah' -Repo 'yq' `
+	-AssetPattern '^yq_linux_amd64$' -BinaryName 'yq' -ArchiveType direct -VersionArguments @('--version') -Linux